+++ ICanPay/Demo/Payment.aspx.cs
@@ -68,7 +68,7 @@ namespace Demo
         /// </summary>
         private void CreateWeChatPayOrder()
         {
-            PaymentSetting paymentSetting = new PaymentSetting(GatewayType.WeChatPay);
+            PaymentSetting paymentSetting = new PaymentSetting(GatewayType.Wechatpay);
             paymentSetting.SetGatewayParameterValue("appid", "wx000000000000000");
             paymentSetting.Merchant.UserName = "000000000000000";
             paymentSetting.Merchant.Key = "000000000000000000000000000000000000000000";
