+++ ICanPay/ICanPay/NotifyProcess.cs
@@ -1,5 +1,4 @@
-﻿using ICanPay.Providers;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml;
@@ -20,7 +19,7 @@ namespace ICanPay
         static string[] yeepayGatewayVerifyParmaNames = { "r0_Cmd", "r1_Code", "r2_TrxId", "r3_Amt", "r4_Cur", "r5_Pid", "r6_Order" };
         static string[] tenpayGatewayVerifyParmaNames = { "trade_mode", "trade_state", "transaction_id", "notify_id", "total_fee", "fee_type" };
         static string[] alipayGatewayVerifyParmaNames = { "notify_type", "notify_id", "notify_time", "sign", "sign_type" };
-        static string[] weixinpayGatewayVerifyParmaNames = { "return_code", "appid", "mch_id", "nonce_str", "result_code" };
+        static string[] wechatpayGatewayVerifyParmaNames = { "return_code", "appid", "mch_id", "nonce_str", "result_code" };
 
         #endregion
 
@@ -38,9 +37,9 @@ namespace ICanPay
                 return new AlipayGateway(gatewayParameterData);
             }
 
-            if (IsWeixinpayGateway(gatewayParameterData))
+            if (IsWechatpayGateway(gatewayParameterData))
             {
-                return new WeChatPayGataway(gatewayParameterData);
+                return new WechatpayGataway(gatewayParameterData);
             }
 
             if (IsTenpayGateway(gatewayParameterData))
@@ -87,9 +86,9 @@ namespace ICanPay
         /// <summary>
         /// 是否是微信支付网关
         /// </summary>
-        private static bool IsWeixinpayGateway(List<GatewayParameter> gatewayParameterData)
+        private static bool IsWechatpayGateway(List<GatewayParameter> gatewayParameterData)
         {
-            return ExistParameter(weixinpayGatewayVerifyParmaNames, gatewayParameterData);
+            return ExistParameter(wechatpayGatewayVerifyParmaNames, gatewayParameterData);
         }
 
 
@@ -126,7 +125,7 @@ namespace ICanPay
             List<GatewayParameter> gatewayParameters = new List<GatewayParameter>();
             ReadQueryString(gatewayParameters);
             ReadForm(gatewayParameters);
-            ReadWeixinpayXml(gatewayParameters);
+            ReadWechatpayXml(gatewayParameters);
 
             return gatewayParameters;
         }
@@ -208,9 +207,9 @@ namespace ICanPay
         /// 读取微信支付的通知
         /// </summary>
         /// <param name="gatewayParameterList">网关通知的参数列表</param>
-        private static void ReadWeixinpayXml(List<GatewayParameter> gatewayParameterList)
+        private static void ReadWechatpayXml(List<GatewayParameter> gatewayParameterList)
         {
-            if (IsWeixinpayNotify())
+            if (IsWechatpayNotify())
             {
                 XmlDocument xmlDocument = new XmlDocument();
                 try
@@ -239,7 +238,7 @@ namespace ICanPay
         /// 是否是微信支付的通知
         /// </summary>
         /// <returns></returns>
-        private static bool IsWeixinpayNotify()
+        private static bool IsWechatpayNotify()
         {
 #if NET35
             string requestType = HttpContext.Current.Request.RequestType;
